Remove pydantic-settings shim
